.ToLE() changed to .value() or removed

+++ rpcs3/Emu/Audio/AudioDumper.h
@@ -67,5 +67,5 @@ public:
 	void WriteHeader();
 	size_t WriteData(const void* buffer, size_t size);
 	void Finalize();
-	const u8 GetCh() const { return (u8)m_header.FMT.NumChannels; }
+	const u16 GetCh() const { return m_header.FMT.NumChannels; }
 };
